加了短信还有修改了好多乱七八糟的东西

+++ ruoyi-yishengxin-back/src/main/java/com/ruoyi/yishengxin/mapper/vipUser/VipTradeHkdBuyMapper.java
@@ -65,5 +65,5 @@ public interface VipTradeHkdBuyMapper {
 	public int deleteVipTradeHkdBuyByIds(String[] ids);
 
 	@Select("select ifNull(sum(buy_number),0) from ysx_vip_trade_hkd_buy where buy_time between #{begin} and #{end}" )
-	int selectSum(@Param("begin") DateTime begin, @Param("end") DateTime end);
+	double selectSum(@Param("begin") DateTime begin, @Param("end") DateTime end);
 }
